requirements.txt
appium-python-client==2.11.1
pytest==7.4.0
selenium==4.15.2
pytest-html==4.1.1
